chore: rename to 'rgbee'
