+++ packages/pages/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "color-demo",
+  "name": "rgbee",
   "private": true,
   "version": "0.0.0",
   "type": "module",
