+++ packages/pages/src/screens/Home.tsx
@@ -1,9 +1,8 @@
 import { useContext } from 'react';
-import { ThemeContext } from '../../../theme/src/theme';
+import { ThemeContext } from '../../../rgbee/src/theme';
 
 const Home = () => {
   const { theme } = useContext(ThemeContext);
-  console.log(theme);
 
   return (
     <pre>{JSON.stringify(theme, undefined, 2)}</pre>
