+++ packages/rgbee/example.md
