+++ packages/rgbee/src/constants.ts
