+++ packages/rgbee/src/defaultThemes.ts
