+++ packages/rgbee/src/utils/__tests__/validation.test.ts
