+++ packages/rgbee/src/utils/calculation.ts
