+++ packages/rgbee/src/utils/conversion.ts
