+++ packages/rgbee/src/utils/validation.ts
